rbfrt
Struct SwitchConnection

pub struct SwitchConnection {
    pub digest_queue: Receiver<Digest>,
    /* private fields */
}

Represents the connection between the switch and the controller.

Example

use rbfrt::SwitchConnection;

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    let switch = SwitchConnection::builder("localhost", 50052)
        .device_id(0)
        .client_id(1)
        .p4_name("my_p4_program")
        .connect()
        .await?;

    Ok(())
}

Fields
digest_queue: Receiver<Digest>

Queue containing all digests of the connected switch.

Implementations

impl SwitchConnection

pub fn builder(ip: &str, port: u16) -> SwitchConnectionBuilder

Creates a new SwitchConnectionBuilder to build the connection between the switch and controller.

The switch is available at the ip address and runs a gRPC server listing on the specified port.

pub async fn execute_operation( &self, request: Request, ) -> Result<(), RBFRTError>

Executes a table operation.

See TableOperation for supported operations, like synchronization of counters or registers.

pub fn has_table(&self, name: &str) -> bool

Checks if the switch has a table with the specified name.

pub async fn get_table_entries( &self, request: Request, ) -> Result<Vec<TableEntry>, RBFRTError>

Retrieves the entries of a single table.

pub async fn get_tables_entries( &self, requests: Vec<Request>, ) -> Result<Vec<TableEntry>, RBFRTError>

Retrieves the entries of multiple tables.

pub async fn write_table_entry( &self, request: Request, ) -> Result<(), RBFRTError>

Writes a single entry into a table.

The entry’s key must not be present in the table to insert the new entry. See update_table_entry to update an existing entry.

pub async fn write_table_entries( &self, requests: Vec<Request>, ) -> Result<(), RBFRTError>

Writes a entries into one or multiple tables.

The entries’ keys must not be present in the table to insert the new entries. See update_table_entries to update existing entries.

pub async fn update_table_entry( &self, request: Request, ) -> Result<(), RBFRTError>

Updates a single entry in a table.

The entry’s key must be present in the table to update the entry. See write_table_entry to insert a new entry.

pub async fn update_table_entries( &self, requests: Vec<Request>, ) -> Result<(), RBFRTError>

Updates multiple entries in one or multiple tables.

The entries’ keys must be present in the tables to update the entries. See write_table_entries to insert new entries.

pub async fn delete_table_entry( &self, request: Request, ) -> Result<(), RBFRTError>

Deletes a entry in a table.

See clear_table to delete all entries inside the table.

pub async fn delete_table_entries( &self, request: Vec<Request>, ) -> Result<(), RBFRTError>

Deletes multiple entries in one or multiple tables.

See clear_tables to delete all entries inside the tables.

pub async fn clear_table(&self, name: &str) -> Result<(), RBFRTError>

Deletes all entries in a table.

See delete_table_entry or delete_table_entries to delete one or multiple entries inside the table.

pub async fn clear_tables(&self, name: Vec<&str>) -> Result<(), RBFRTError>

Deletes all entries in multiple tables.

See delete_table_entry or delete_table_entries to delete one or multiple entries inside the tables.

pub async fn get_register_entry( &self, request: Request, ) -> Result<Register, RBFRTError>

Reads the value of a register.

pub async fn get_register_entries( &self, requests: Vec<Request>, ) -> Result<Register, RBFRTError>

Reads the values of multiple registers or indices.

pub async fn write_register_entry( &self, request: Request, ) -> Result<(), RBFRTError>

Writes a value into a register.

pub async fn write_register_entries( &self, requests: Vec<Request>, ) -> Result<(), RBFRTError>

Writes values into multiple registers or indices.