rbfrt::table

Struct TableEntry

pub struct TableEntry {
    pub table_id: u32,
    pub table_name: String,
    pub match_keys: HashMap<String, MatchValue>,
    pub default_entry: bool,
    pub action: String,
    pub action_data: Vec<ActionData>,
}

Represents a table entry.

Fields
table_id: u32

Id of the table.

table_name: String

Name of the table.

match_keys: HashMap<String, MatchValue>

Names and their values of the match keys.

default_entry: bool

Flag indicating if this is the default entry for the table.

action: String

Name of the associated action

action_data: Vec<ActionData>

Action data of the action, empty if not parameters are provided.

Implementations

impl TableEntry

pub fn get_key(&self, name: &str) -> Result<&MatchValue, RBFRTError>

Returns the MatchValue of the match key with the given name.

pub fn has_key(&self, name: &str) -> bool

Returns whether a match key with the given name is present.

pub fn get_action_data(&self, name: &str) -> Result<&ActionData, RBFRTError>

Returns the ActionData which key has the given name.

pub fn has_action_data(&self, name: &str) -> bool

Returns whether an action data has a key with the given name.

pub fn get_action_name(&self) -> &str

Returns the action’s name.