rbfrt::util
Struct Port

pub struct Port { /* private fields */ }

Represents a port of the switch.

Implementations

impl Port

pub fn new(port: u32, channel: u8) -> Port

Creates a new Port with the provided frontpanel port number and channel. The speed is 1Gb/s, the auto negotiation is default, the forward error correction is disabled, and the loopback is disabled.

pub fn speed(self, speed: Speed) -> Port

Sets the speed.

pub fn loopback(self, loopback: Loopback) -> Port

Sets the loopback mode.

pub fn enable(self) -> Port

Enables the Port.

pub fn disable(self) -> Port

Disables the Port.

pub fn auto_negotiation(self, auto_neg: AutoNegotiation) -> Port

Sets the auto negotiation.

pub fn fec(self, fec: FEC) -> Port

Sets the forward error correction.

pub fn get_speed(&self) -> &Speed

Returns the configured speed.

pub fn get_loopback(&self) -> &Loopback

Returns the configured loopback mode.

pub fn get_enabled(self) -> bool

Returns if the Port is enabled.

pub fn get_auto_negotiation(&self) -> &AutoNegotiation

Returns the configured auto negotiation.

pub fn get_fec(&self) -> &FEC

Returns the configured forward error correction.

pub fn get_dev_port(&self) -> Option<u32>

Returns the dev_port of the frontpanel port number.

pub fn get_frontpanel_port(&self) -> (u32, u8)

Returns a tuple (port, channel) of the frontpanel port.